Speedup completions
Don't spend too much time in Jedi. It can be slow.
Return what we've got so far as fast as we can.

the timeout of 0.4 second was chosen by trial and error
with:

    import matplotlib
    import matplotlib.patches

    matplotblib.<tab>

keeping in mind that I have a fast machine.
There is an overhead associated with rerunning the completion so a too
short value, never leaves Jedi the chance to progress in computing more
completions.