Backport PR #10260: Fix the rest of the colors used on Windows so they are more visible
This applies a hack similar to the one in
7b7d560822d4e9b6ab486393e22e305c502ff6d1 to change the prompt colors
and exception colors so they are more visible as shown in issue  10238

If this fix is acceptable it would be nice if this is backported to IPython 5 as this is very visible. Every user I see using IPython on Windows sees this and most don't seem to understand that they can change it. And on top of that this is especially annoying when using IPython on a computer you don't own or a VM. Having to reconfigure on each one...

Fixes:  10238