Minor improvements to completion
If object being completed is a function, append `()` to match the
convention used in cpython doc. This include methods. Append `()` only
in the _displayed_ menu, does not insert the ().

If completion is relatively to long (mostly if is does not come from
jedi), and use a fully qualified name with more than 3 parts, ellides
the middles one to `…`. (This closes #9835).