Fix virtualenv warning for Cygwin
Getting `Attempting to work in a virtualenv...` warnings when running in Cygwin and starting ipython in a virtualenv. (Also, using a virtualenvwrapper, not sure if it makes a difference)

Turns out `p_venv` looks like this `\cygdrive\c\users\igonato\envs\myenv` when `sys.executable` is like this `c:\users\igonato\envs\myenv\scripts\python.exe`.

This should make the check a bit more robust when running in Cygwin

Cheers!