Show signature with Jedi.
When completion are function, jedi is capable of giving us the function
signature. This adds a signature field to the (still unstable) completer
API and make use of it in the IPython terminal UI.

It is not (yet) exposed by the ipykernel.

Additionally add typing to a couple of locations.