[Question] shadow namespace
We do a lot of dance with user_ns and our injected variable.
In particular _,__,___ and rotate these while still making sure the user
can overwrite and if _ is user defined not push this one to __, but the
actual older value.

It seem to me that ChainMap would be a way cleaner way of doing this,
with a slightly different semantic as you can't delete things from the
shadow namespace directly.

Do anybody see an obvious reason not to use ChainMap ? If so I might
experiment with it.