Add metadata to DisplayObject (#10614)
The aim of this PR is to get it so that DisplayObject has a metadata API that can be inherited by it's subclasses.

One of the challenges in designing this is to not destroy those subclasses that had already defined their own metadata API (such as Image). 

This also adds a way for SVG objects to surface metadata when returned or displayed.