Deactivate sorting on dict keys when pretty printing.
Closes #10110

That is a backward incompatible change, but as the dictionary are (as an
implementation details of Python as of 3.6) ordered and teaching Python
often need to show that (for kwargs), it is (IMHO) a better choice to
remove the sorting.

If one id unhappy with that it is still possible to register its own
pretty printer.