Fix jedi 0.11 – be optimistic.
Instead of not trying Jedi completion if Jedi parsing of wether we are
in a string fails assume that we should alway try jedi, and do not only
if jedi manage to prove we are currently in a string.

This should give some false positive (like triggering jedi in some
unfinished strings), but no false negative – like for example with Jedi
0.11 which is just out.

In the end we should do a proper fix; but I'd like to get that soon as
otherwise Jedi is quasi deactivated when version 0.11 is installed
(released a few days ago).