Silence tokenization errors in ultratb.
Currently, the ultratb module attempts to tokenize the code it finds in
tracebacks, and issues a scary-sounding warnings about potentially invalid
tracebacks if it fails. However, this tokenization is known to fail in some
cases, notably if the reported error line is the terminating line of a
multiline, triple-quoted string literal.

This change fixes the issue by switching this to a `logging.debug` message,
which silences it by default without completely removing it for debugging
purposes.

Fixes #6864.