Change return type in docstring to literal tuple to avoid bogus cross-ref.
Before this change the Sphinx tried to produce cross-reference links for
the returned (stdout, stderr) tuple.
The docs build produced a warning because there are 2 possible cross-ref
objects for `stdout`,
one in `IPython.testing.iptestcontroller.TestController`,
the other in `IPython.utils.capture.CapturedIO`.
Neither of those is correct in the context of the `ipexec()` function.
Changing literal ``(stdout, stderr)`` avoids the warning by preventing
the attempt to produce cross-references.