Detect string completion when no syntax error
Previously the behaviour of completing would be different for `d['`
compared to `d['']`, because disabling of `jedi` relied on detecting a
syntax error, which isn't present in the latter. With this fix, this is
correctly detected.

Unfortunately, the change is very difficult to test, because there is
no easy way to test for _absence_ of any `jedi` completions.