Refactor of coloring and traceback mechanism.
This should not change behavior be should slightly clean the code to be
a bit more pythonic, and start to introduce methods that avoid side
effect (that is to say return values instead of directly writing to
stdout). I keep it relatively small and simple for now.

My goal is to slowly change the traceback mechanisme to separate buiding
the data structure from rendering it to potentially introduce richer
tracebacks.  Locally I've also started from the other side (re build a
traceback rendered from scratch) to see what's needed.