Fix `up` through generators in postmortem debugger
By passing the lowest traceback element into the debugger, we require it
to use frame.f_back to find older frames. However, f_back is always None
for generator frames. By providing a higher-up traceback element, pdb
can traverse down the traceback.tb_next links, which do work correctly
across generator calls.

Furthermore, pdb will not do the right thing if it is provided with a
frame/traceback pair that do not correspond to each other - it will
duplicate parts of the callstack. Instead, we provide None as a frame,
which will cause the debugger to start at the deepest frame, which is
the desired behavior here.