Fix magic directive and role.
We were directly modifying state instead of using blessed API (also some
api will change so be future proof with sphinx master).

This should restore some ability of the :magic: and :cellmagic:
directive to both support link to magics name when the text in the
directive contains (or not) leading % and %%, it will also automatically
prepend the right number of % for magics and cell magics.

A couple of other related fixes.