Fix to allow entering docstring into IPython.
The EscapeTransformer find method was assuming incorrectly that every
line would end with either a NEWLINE or EOF, while this is not the case
when encountering multiple line string.

This fixes that by making sure we don't index outside of bounds.

With this IPython will correctly add a newline at the CLI.

Closes #11391