Fix miss-capturing of assign statement after a dedent.
closes #11415

This fixes a bug where assign statement were miscaptured when occuring
after a dedent. This was due to the fact that :

>>> '' in '({['
True

That is to say the empty string is in any strings.
Add a couple of integration tests and unit tests as well, and also add a
warning to public function when not used properly, in particular, check
that lines passed to make_tokens_by_line do end with an endline marker
(at least for the first line), otherwise the function does not behave
properly.