Remove leading empty lines
When running a cell, the leading indentation on the first line is
removed from all of the cell's lines. This causes indentation errors
on subsequent lines in case the first line consists entirely of
whitespace matching a valid level of indentation that is greater than
the least indented line.

This change avoids the issue by first removing any leading,
whitespace-only lines.

Closes #11594