Prepare some test for pytest in completers.
swap the context manager and the function definition.
It is undefined wither the test are collected and then run at once; or
ran while they are being yielded.

In the fist case (what pytest does) the function was called outside the
context manager, in the later case (node) it was working as expected.