Try to speedup autoreload by walking all the object using gc.
Unsure if this work; we were avoid this initially as this crawl more
objects; but at the same time we do less objects introspection.

Another possibility is to look for all gc.referers(OldClass), as I
believe their __class__ will point to that; so they should be related by
only one degree.

May Close #11817