Adding VI-mode in prompt
From the code it is apparent that if vi-mode is enabled, it should
display in the prompt. However, as the compare done is done against an
Enum class, it will always return False.

The solution is either to access the Enum's `.value` attribute, or to
compare it to the declared Enum value directly. The latter is likely
more readable.