Support vi-mode 'InputFormat' prompt
Prompt defining what vi-mode is in, is either 'vi-insert' or 'vi-navigate' (or something 'vi-...' I imagine).
From this `[3:6]` is extracted to get either 'ins' or 'nav' in the prompt.
This is consistent with the behavior I experience when manipulating a custom prompt from the most updated pip installable package.

However, in the case of installing the current master from source, the prompt is on the format 'InputFormat.INSERT' and 'InputFormat.NAVIGATION'.