Improves detection of whether tab-completion is in a string and supresses Jedi.
Refs #10926 and #11530

Jedi results swamp file_matches and dict_key_matches in
tab-completion, which is a real nuisance.  The logic in the jedi
completor tried to catch cases where it was "in a string", but that
logic only looked at the previous character, and was a little fragile,
breaking in lots of cases such as:

'./<tab>
"mypath/<tab>

etc.

This seems a bit more robust in that it searchs for the first token in
the current parser tree and checks if its value starts with ' or ".

This detection of "in a string" then turns of jedi and returns some
sanity to the set of matches.