WIP: Return proxy object from magic_arguments.
This changes the @magic_arguments decorator to return a proxy wrapper around
the decorated function instead of repeatedly mutating the underlying function
in place.

Based on Twitter conversation with @Carreau here:
https://twitter.com/scottbsanderson/status/1260280451710795777.

**NOTE**: A bunch of the changes here are made solely to appease the test suite
of test_magic_arguments, which makes a bunch of assertions about attributes
that do or don't get set on the decorated function (e.g., it asserts that
has_arguments gets set, and that argcmd_name sometimes gets set and sometimes
doesn't). I don't think any external code actually depends on these behaviors,
so the code could almost certainly be cleaned up after changing the tests to
not depend on implementation details. I chose to defer that change in this
commit so that there's a clean separation between internal refactoring and
behavior changes.