Fix auto formatting to be called only once.
There seem to be a small race-condition when trying to reformat the text
twice in a short amount of time where some of the non-changed test
diapear. For example

    async def b():pass

Lead to two reformatting where one does not see the `async` keyword
anymore., and would thus dedent the def by 1 space.