Show make sure to flush hidden exception to screen.
This is really unlikely but can happen if the exception is raised from
within a stack of hidden frame, in which case we do want to say there
were some hidden frames.

I guess in that case we actually really want to not hide that last one,
but this might really be a rare case.

This is already fixed in the backport to 7.x