Allow passing extra options to TerminalPdb.pt_init
Currently, subclasses of TerminalPdb (like in https://github.com/kmaork/madbg/blob/master/madbg/debugger.py#L55) will have to duplicate the entire pt_init method in order to add additional arguments for the prompt session (which accepts about 40 of them). An example would be a subclass that wants to write the prompts to a different output stream than the default. After this change it would just have to override the call to pt_init with `pt_init(dict(input=input, output=output))`