Move Extraction of local scope to a method
When calling ipython magic from pdb, the locals need to be set from the pdb locals, which is different from the frame locals. Moving this call to a method enables a subclass to change the behaviour of `get_local_scope` to make this possible. This would enable using `%timeit` while debugging in Spyder for example