Don't return the last line in `find_user_code`
Changes `InteractiveShell.extract_input_lines` such that if the empty
histrange is given, the last line of history is omitted.

The reason for this is that `InteractiveShell.find_user_code`, the only
place where the aforementioned method is used, is only called by magics,
where the empty histrange argument would otherwise include the magics
themselves.