Add history file to debugger.
This adds a configurable
`InteractiveShell.debugger_history_file=...` which default to
`~/.pdbhistory`, that store what is typed in ipdb;
this make it easy to persist across sessions.

Some of the logic is moved into the debugger itself so that existance
and creation of file is used only once Pdb is started.