use sphinxify (if available) in object_inspect_mime path
Background: if we use `%pinfo` line magic (or equivalently `?`; and similarly
for `pinfo2` and `??`), we will call `interactiveshell`'s `_inspect` method
which automatically uses the docrepr module (if available and activated) to
beautifully format docstrings.

However, if we use the lower-level `object_inspect_mime` public method in
`interactiveshell`, docrepr is not used.

(Sidenote: practically speaking, this function is called from JupyterLab's
Context Help, which creates an "inspect_request" message, which in turns calls
`object_inspect_mime`.)

This commit enables `object_inspect_mime` to use the same functionality as
`_inspect`.