Require pygments>=2.4.0
As noted in #13441, running ipython with an old version of pygments
leads to problems. Ipython sends ANSI color names to pygments to color
output, but these names aren't in old versions of pygments.

Before: with pygments 2.3.1 and ipython 8.0.0, opening an ipython
instance and running

    In [1]: 1 / 0   # Expect ZeroDivisionError

will crash ipython as `ansiyellow` is used to highlight the error.

This PR requires pygments>=2.4.0, which is when pygments changed their
ANSI color names.