Support square bracket indexing with literal with pinfo/?/pinfo2/??
We add limited support for evaluating `[xxx]` where xxx is a numeric
literal. This still abit dangerous as __getindex__ can run arbitrary
code.

We could support a wider range of constructs, but this would start to
reach into complex and dangerous territory. Alredy this with lead to
weird suff like
   >>> def foo(*arg, **kwargs):pass
   >>> c = [foo]
   >>> c[0]?
   Signature: c[0](*args, **kwargs)

Instead of

   >>> c[0]?
   Signature: foo(*args, **kwargs)

As the name of the object is used in the signature is the one if the
main namespace.

This should fix https://github.com/ipython/ipython/issues/13780