FEAT: Fast TB.
Try to detect when one of the file we are trying to highlight is too
big, and fallback on pre-8.0 traceback code, that avoids stack_data.

You can configure the limit with:

>>> from IPython.core import ultratb
>>> ultratb.FAST_THRESHOLD = 200

We are trying to _guess_ whether the traceback we will render is in a
file that will require to parse more than FAST_THRESHOLD lines, though
it is actually difficult to get this value correctly, so in some case,
the slow path may be used, even if files are longer than FAST_THRESHOLD,
especially if the crashing code is at the beginning of the file.