Make oinspect._render_signature code more closely align with the upstream code it is modifying for easier maintenance
oinspect._render_signature is basically a copy of inspect.Signature.__str__ with a few tweaks. However, comparing the two functions is hard since it was basically rewritten for _render_signature. This commit makes _render_signature nearly identical to inspect.Signature.__str__ so it can be easily compared in the future, with clear notes about what was changed.

There should be no behavior difference before and after this change.