Refactor Inspector._get_info to make it easier to subclass and modify behavior.
This factors out the logic to append an info field to the inspect_reply mimebundle and the logic for what information is in the mimebundle into separate functions that can easily be overridden by an inspector subclass. This allows a subclass to easily:

* format information into yet another mimetype besides text/plain and text/html
* modify or add to the default information

without having to copy the default implementation.