docs: add more pointers to configuration of embedded IPython
The "Running IPython from Python" section of the docs under "Configuring
IPython" gives a nice example of how to set IPython configuration
options when embedding IPython. Add more pointer to this part of the
documentation by linking to it from the docstrings of
IPython.{embed_kernel, start_ipython, start_kernel} and
IPython.terminal.embed.embed.

While at it, consistenly refer to the config argument as "a traitlets
:class:`Config` object".

Finally, also mention that options can be set with a 'Config' object in
the small paragraph under "IPython options".