MAINT: reintroduce a OInfo get method for backward compatibility.
Reintroduce a OInfo get method for backward compatibility.

OInfo was converted to a dataclass from dict in commit 0fef298ad70ff72,
but some external code is still be using it.

This commit reintroduces the method, but we don't deprecated it yet !

Fixes https://github.com/h5py/h5py/issues/2253