Provide and easier way to generate magics and pre-post hooks
This introduce a code base way of modifying the ast.
This allow to use Template strings with the two special variable names
names `__code__` and `__ret__` allowing to quickly write a magic, or
hooks that modify the ast.

This also introduce a `%code_wrap` cell magic to on the fly wrap code.
It is this easy to for example modify IPython to say time each block of
code, or retry them, or wrap them in try/except and analyse the error
message, profile...

Note that this is not new, but simply convenience function and
utilities, especially around hygiene.