Alternative stricter fix: explicit isoformat(" "),
I think we need to avoid registering a default adapter, especially if
CPython want to remove them.

So we are going to call .isoformat(" ") explicitly. For backward compat
we'll keep the delimiter as " ", but we might want to consider using the
default of T later on, but still keeping in mind that old history have a
space.