Fix many py2-style prints in docs and commented code.
The main interest was to transform py2-style print statements that still
appeared in some docstrings (e.g. for %history) and would thus show up
when looking at the docs either interactively or in the rendered
website.  I also fixed a number of commented-out print statements at the
same time, mostly because it was easier for me to just regexp-change all
of them so that I could focus on the others.