Massive colors and themeing refactor
This massively rewrite the color and theme handling.

Instead of messing with colors internally, we try – as much as possible,
to yield pairs of (token types, values) and feed that to a pygments
formatter.

I'm going to assume this is going to break stuff, hence why I want it
early in the 9.0 release cycle.