Allow multiline autosuggest and draft of LLM completion.
This is a preview-draft of modification to IPython in order
to allow multi-line suggestion in the terminal, and add the necesarry
hooks and examples, in order to trigger LLM autocomplation via a
shortcut.

This will not be complete, but will be available as a preview in IPython
8.32 in order to gather feedback.

This tries to be as compatible as possible with jupyter_ai and
jupyter_ai_magics, except the Provider cannot take parameters in their
constructor.