Try to fix history manager thread leakage
Right now the cleanup was mostly done at process exit time,
but this leads to issues in downstream test that start to have dozen
of threads blocking, and make finding deadlocks harder

This PR try to cleanup the cleaning logic.

 - Try to seprate saving thread from history manager
 - Make sure to not keep reference to history manager in multiple places.
 - att a utility to refuse to create more then N
   concurrent history manager HistorySavingThread
   instances.
   - default to 1 in conftest.py
   - allow to set it to N via pytest fixture.

Some of this should likely be either context managers, or finalisers.