Try to close more resouce on windows in system command.
There is a resource leak spotted in ipython/ipykernel#1291, that only
affects windows, this try to properly close the BufferReader/writrers.

I find it weird to have to close things we did not set on the
subprocess, I guess this is because it is supposed ot be used as a
context manager, unfortunately the CM does .wait(), and we don't want to
wait here.